move get* data functions into node_data module
